+++ fortpy/isense/classes.py
@@ -1,49 +1,85 @@
 from . import cache
 import re
 from fortpy.elements import Function, Subroutine, ValueElement, Executable, TypeExecutable
+from fortpy.printing import docs
 
 class BaseDefinition(object):
-    def __init__(self, context, start_pos):
-        self._start_pos = start_pos
+    def __init__(self, context, element):
         self._context = context     
+        self._element = element
+
+    @property
+    def name(self):
+        """Returns the name of the underlying code element."""
+        return self._element.name
 
     @property
     def type(self):
         """
         The type of the definition. One of the classes in elements.py.
         """
-        return type(self._context.element)
+        return type(self._element).__name__
 
-    def _path(self):
+    @property
+    def module_path(self):
         """The module path."""
-        return self._context.module.filepath
+        return self._element.module.filepath
 
     @property
     def module_name(self):
         """
         The module name
         """
-        return str(self._context.modulename)
+        return str(self._element.module.name)
 
     @property
     def line(self):
         """The line where the definition occurs (starting with 0)."""
-        return self._start_pos[0]
+        return self._element.module.linenum(self._element.start)
 
     @property
     def column(self):
         """The column where the definition occurs (starting with 0)."""
-        return self._start_pos[1]
+        line = self.line
+        return self._element.start - self._element.module.charindex(line, 0)
 
+    @property
     def docstring(self):
-        """Return a document string list for this completion object.
+        """Returns the <summary> tag of the code element represented by this
+        completion.
         """
-        return self._context.element.docstring
+        if hasattr(self._element, "summary"):
+            return self._element.summary
+        else:
+            return "Undocumented Symbol"
+
+    @property
+    def fulldoc(self):
+        """Returns the full docstring for the code element."""
+        return docs.format(self._element)
 
     @property
     def description(self):
-        """A textual description of the object."""
-        return self._context.element.summary
+        """Returns the full docstring information for the element suggested 
+        as a completion."""
+        result = ""
+        if isinstance(self._element, ValueElement):
+            if self._element.kind is not None:
+                result = "{}({}) | {}".format(self._element.dtype, self._element.kind,
+                                          self._element.summary)
+            else:
+                result = "{} | {}".format(self._element.dtype,
+                                          self._element.summary)
+        elif isinstance(self._element, Executable):
+            result = "({})".format(self._element.parameters_as_string())
+        elif isinstance(self._element, str):
+            result = "Intrinsic Fortran Symbol"
+        elif isinstance(self._element, TypeExecutable):
+            result = self._type_description()
+
+        #Clean off any line breaks from the XML and excessive whitespace.
+        cleaned = re.sub("\s+", " ", result.replace("\n", " "))
+        return cleaned
 
     @property
     def full_name(self):
@@ -71,10 +107,7 @@ class Completion(BaseDefinition):
     provide additional information about a completion.
     """
     def __init__(self, context, element, like_name_length):
-        super(Completion, self).__init__(context, context.pos)
-
-        self._context = context
-        self._element = element
+        super(Completion, self).__init__(context, element)
         self._like_name_length = like_name_length
 
     def _complete(self, like_name):
@@ -119,43 +152,6 @@ class Completion(BaseDefinition):
     def __repr__(self):
         return '<%s: %s>' % (type(self).__name__, self.name)
 
-    @property
-    def docstring(self):
-        """
-        :param fast: Don't follow imports that are only one level deep like
-            ``import foo``, but follow ``from foo import bar``. This makes
-            sense for speed reasons. Completing `import a` is slow if you use
-            the ``foo.documentation(fast=False)`` on every object, because it
-            parses all libraries starting with ``a``.
-        """
-        if hasattr(self._element, "summary"):
-            return self._element.summary
-        else:
-            return "Intrinsic Fortran Symbol"
-
-    @property
-    def description(self):
-        """Returns the full docstring information for the element suggested 
-        as a completion."""
-        result = ""
-        if isinstance(self._element, ValueElement):
-            if self._element.kind is not None:
-                result = "{}({}) | {}".format(self._element.dtype, self._element.kind,
-                                          self._element.summary)
-            else:
-                result = "{} | {}".format(self._element.dtype,
-                                          self._element.summary)
-        elif isinstance(self._element, Executable):
-            result = "({})".format(self._element.parameters_as_string())
-        elif isinstance(self._element, str):
-            result = "Intrinsic Fortran Symbol"
-        elif isinstance(self._element, TypeExecutable):
-            result = self._type_description()
-
-        #Clean off any line breaks from the XML and excessive whitespace.
-        cleaned = re.sub("\s+", " ", result.replace("\n", " "))
-        return cleaned
-        
     def _type_description(self):
         """Gets the completion description for a TypeExecutable."""
         #This is a little tricker because the docstring is housed